Add support for boolean parameters in backend_info
Introduced in Asciidoctor [1.5.7], the :supports_templates option allows
using custom converters with custom templates using the `:template_dirs`
option or the `-T` command line option.

To enable it in a converter, it needs to call the `supports_templates`
method in its initializer without any arguments.

This patch adds the ability to set a backend_info parameter to `true`,
which will add a call to that function without any arguments.

[1.5.7]: https://github.com/asciidoctor/asciidoctor/commit/ec8e7e002be41a12f79c2e2629ac7ea15103438a